moved EqSearch to new searches directory
