+++ src/main/java/hello/controllers/EarthquakesController.java
@@ -1,7 +1,7 @@
 package hello.controllers;
 
-import hello.EarthquakeQueryService;
-import hello.EqSearch;
+import hello.services.EarthquakeQueryService;
+import hello.searches.EqSearch;
 import hello.services.EarthquakeQueryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
